refactor(payment): 금액 자료형 BigDecimal -> Long 변경

+++ payment-service/src/main/java/com/rushcrew/payment_service/presentation/dto/request/PaymentRequest.java
@@ -10,7 +10,7 @@ public record PaymentRequest(
         @NotNull(message = "주문 ID는 필수입니다.")
         UUID orderId,
         @NotNull(message = "결제 금액은 필수입니다.")
-        BigDecimal totalAmount
+        Long totalAmount
 ) {
         public PaymentCommand toCommand() {
                 return new PaymentCommand(
